+++ src/sequence.rs
@@ -4,20 +4,20 @@ use skia_safe::{surfaces, Color, ISize};
 use vector2d::Vector2D;
 
 pub struct Sequence{
-    start_frame: u32,
-    end_frame: u32,
-    resolution: Vector2D<i32>,
+    start_frame: usize,
+    end_frame: usize,
+    resolution: Vector2D<usize>,
     info: skia_safe::ImageInfo,
     elements: RefCell<Vec<Box<dyn elements::Element>>>,
 }
 
 impl Sequence {
-    pub fn new(width: i32, height: i32) -> Sequence {
+    pub fn new(width: usize, height: usize) -> Sequence {
         Sequence {
             start_frame: 0,
             end_frame: 0,
             resolution: Vector2D::new(width, height),
-            info: skia_safe::ImageInfo::new_a8(skia_safe::ISize::new(width, height)),
+            info: skia_safe::ImageInfo::new_a8(skia_safe::ISize::new(width as i32, height  as i32)),
             elements: RefCell::new(Vec::new()),
         }
     }
@@ -26,8 +26,8 @@ impl Sequence {
         self.elements.borrow_mut().push(e)
     }
 
-    pub fn render_frame<'a>(&mut self, frame: u32) -> Vec<u8> {
-        let size = ISize::new(self.resolution.x, self.resolution.y);
+    pub fn render_frame<'a>(&mut self, frame: usize) -> Vec<u8> {
+        let size = ISize::new(self.resolution.x as i32, self.resolution.y as i32);
         let mut surface = surfaces::raster_n32_premul(size).expect("surface");
         let mut canvas = surface.canvas();
